Treat gmock-matchers.h as a system header.
Clang warns about the comma operator usage in this file because it
looks like the author might have meant to use the value of the first
part of the comma operator, so it warns that it isn't being used.
Treat this header as a system header to shut up warnings from it
until we can resolve this with upstream.

Test: make checkbuild
Bug: None
Change-Id: I74d5bf5d469b80812d72ada38fbfa794295f87a7